+++ packages/extension/src/shared/multisig/multisig.service.ts
@@ -1,3 +1,5 @@
+import urlJoin from "url-join"
+
 import { ARGENT_MULTISIG_URL } from "../api/constants"
 import { Fetcher, fetcher } from "../api/fetcher"
 import { Network } from "../network"
@@ -38,4 +40,35 @@ export async function fetchMultisigDataForSigner({
   })
 
   return ApiMultisigDataForSignerSchema.parse(data)
+}
+
+export const getMultisigAccountData = async ({
+  address,
+  networkId,
+}: {
+  address: string
+  networkId: string
+}) => {
+  try {
+    if (!ARGENT_MULTISIG_URL) {
+      throw new Error("Multisig endpoint is not defined")
+    }
+    const url = urlJoin(ARGENT_MULTISIG_URL, `${networkId}/${address}`)
+    return fetcher<{
+      content: {
+        address: string
+        creator: string
+        signers: string[]
+        threshold: number
+      }
+    }>(url, {
+      method: "GET",
+      headers: {
+        Accept: "application/json",
+        "Content-Type": "application/json",
+      },
+    })
+  } catch (e) {
+    throw new Error(`An error occured ${e}`)
+  }
 }
